+++ docs/versionhistory.rst
@@ -5,6 +5,11 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 **UNRELEASED**
 
+- **BACKWARD INCOMPATIBLE** Bumped minimum Asphalt version to 4.8
+- **BACKWARD INCOMPATIBLE** Refactored component to only provide a single renderer
+  (you will have to add two components to get two renderers)
+- **BACKWARD INCOMPATIBLE** Dropped the context attribute (use dependency injection
+  instead)
 - Dropped explicit run-time type checking
 
 **3.0.2** (2022-04-14)